test(e2e): wire e2e suite into npm test, fix and prune stale cases
npm test ran only 2 of the test files; the 6-suite e2e runner was never
invoked, and two suites had gone stale. Make the suite meaningful and wire it
in so `npm test` is trustworthy for new contributors.

- npm test now runs sort + auth + test/e2e/run.js (add test:e2e script)
- schema F2-TC8: rewrite to the current `compound` rule type; the old
  `specific_sets` rule was removed, so the test silently passed nothing
- delete ocr.test.js: tested a removed OCR feature via a test-local regex and
  tautological source-greps (asserted nothing real)
- repurpose rebrand.test.js -> server.test.js: drop one-time rebrand greps,
  keep the only full server-boot + health/auth smoke test
- README: note test/ now covers unit + e2e
