+++ backend/package.json
@@ -6,7 +6,8 @@
   "scripts": {
     "start": "node src/server.js",
     "dev": "nodemon src/server.js",
-    "test": "node test/sort.test.js && node test/auth.test.js"
+    "test": "node test/sort.test.js && node test/auth.test.js && node test/e2e/run.js",
+    "test:e2e": "node test/e2e/run.js"
   },
   "dependencies": {
     "@huggingface/transformers": "^4.2.0",
